new blog post & small change to index.html
The blog post is just an update and not a big post on a separate page. As for the changes to the index.html page, I originally had a warning that my website wasn't responsive/mobile-friendly. However, since I _have_ implemented responsive CSS to my website, that warning has changed.